+++ src/Type/Definition/Description.php
@@ -1,6 +1,6 @@
 <?php declare(strict_types=1);
 
-namespace GraphQL\Utils;
+namespace GraphQL\Type\Definition;
 
 use Attribute;